clean macros
